fix: Fix detection of fd leaks
For test cmds that start multiple child processes, the outputs of valgrind for each process can be interleaved.
The previous `awk` command to detect open file descriptor errors from the valgrind output could not handle this.

Now, the `awk` command considers the pids of the valgrind output to fix this.