fix: Fix too high resource usage of valgrind
If the user has modified their PATH env variable, so that the path of f.e. `ls` is different than expected, valgrind will follow those child processes.
In test cases with a lot of child processes, this can lead to lockups because of too high CPU and RAM usage.

A common example is when the user installs capt.